minor fixes

+++ src/kedro_graphql/pipelines/example00/nodes.py
@@ -5,5 +5,5 @@ generated using Kedro 0.18.4
 import time
 
 def echo(text: str, example: str, parameters: dict) -> str:
-    time.sleep(int(parameters.get("duration", 1)))
+    time.sleep(float(parameters.get("duration", 1)))
     return text
